Provide more descriptive validation errors
tv4 (used for response validation) generally provides a generic error message such as "Data does not match any schemas from "anyOf"". It also provides an array of sub-errors which include any validation issues. While some of these issues may be ignored, it can also include the exact reason for the validation failure (such as a missing field). Include this in the validation error message to aid in troubleshooting.